Split socket in context to hook
