+++ src/features/call/components/common/modal/modal-add-user.tsx
@@ -5,9 +5,9 @@ import { cn } from '@/utils/cn';
 import { useAuthStore } from '@/stores/auth.store';
 import { SelectedList } from '@/features/chat/rooms/components/selected-list';
 import { SearchList } from '@/features/chat/rooms/components/search-list';
-import { useVideoCallStore } from '../../store/video-call.store';
+import { useVideoCallStore } from '../../../store/video-call.store';
 import { getRoomService } from '@/services/room.service';
-import { useParticipantVideoCallStore } from '../../store/participant.store';
+import { useParticipantVideoCallStore } from '../../../store/participant.store';
 import { SOCKET_CONFIG } from '@/configs/socket';
 import socket from '@/lib/socket-io';
 import { SearchInput } from '@/components/data-entry';
