+++ src/features/call/components/common/modal/modal-request-join-room.tsx
@@ -2,8 +2,8 @@
 import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from '@/components/feedback';
 import socket from '@/lib/socket-io';
 import { SOCKET_CONFIG } from '@/configs/socket';
-import { useVideoCallStore } from '../../store/video-call.store';
-import { useParticipantVideoCallStore } from '../../store/participant.store';
+import { useVideoCallStore } from '../../../store/video-call.store';
+import { useParticipantVideoCallStore } from '../../../store/participant.store';
 
 export const RequestJoinRoomModal = () => {
     const { room } = useVideoCallStore();
