+++ src/features/call/components/common/modal/modal-stop-doodle.tsx
@@ -2,7 +2,7 @@
 import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from '@/components/feedback';
 import socket from '@/lib/socket-io';
 import { SOCKET_CONFIG } from '@/configs/socket';
-import { useVideoCallStore } from '../../store/video-call.store';
+import { useVideoCallStore } from '../../../store/video-call.store';
 import { useAuthStore } from '@/stores/auth.store';
 
 export const ConfirmStopDoodle = () => {
