+++ src/features/call/interfaces/socket/doodle.interface.ts
@@ -1,0 +1,4 @@
+export interface IStartDoodlePayload {
+    image_url: string;
+    name: string;
+}
