+++ src/features/call/interfaces/socket/join.interface.ts
@@ -1,0 +1,8 @@
+import SimplePeer from "simple-peer";
+
+export interface IJoinCallPayload {
+    signal: SimplePeer.SignalData;
+    callerId: string;
+    user: any;
+    isShareScreen: boolean;
+}
