+++ src/features/call/interfaces/socket/signal.interface.ts
@@ -1,0 +1,6 @@
+export interface IReturnSignal {
+    id: string;
+    signal: string;
+    user: any;
+    isShareScreen: boolean;
+}
