+++ src/features/call/store/participant.store.ts
@@ -1,15 +1,15 @@
 import { create } from 'zustand';
-import ParicipantInVideoCall from '../interfaces/participant';
+import ParticipantInVideoCall from '../interfaces/participant';
 
 export type VideoCallState = {
-    participants: ParicipantInVideoCall[];
+    participants: ParticipantInVideoCall[];
     usersRequestJoinRoom: any[];
     peerShareScreen: any[];
     addPeerShareScreen: (peer: any) => void;
     removePeerShareScreen: (id: string) => void;
     clearPeerShareScreen: () => void;
-    updateParticipant: (participants: ParicipantInVideoCall[]) => void;
-    addParticipant: (participant: ParicipantInVideoCall) => void;
+    updateParticipant: (participants: ParticipantInVideoCall[]) => void;
+    addParticipant: (participant: ParticipantInVideoCall) => void;
     setStreamForParticipant: (stream: MediaStream, socketId: string, isShareScreen: boolean) => void;
     removeParticipant: (socketId: string) => void;
     removeParticipantShareScreen: (socketId: string) => void;
@@ -26,10 +26,10 @@ export const useParticipantVideoCallStore = create<VideoCallState>()((set) => ({
     participants: [],
     usersRequestJoinRoom: [],
     peerShareScreen: [],
-    updateParticipant: (participants: ParicipantInVideoCall[]) => {
+    updateParticipant: (participants: ParticipantInVideoCall[]) => {
         set(() => ({ participants }));
     },
-    addParticipant: (participant: ParicipantInVideoCall) => {
+    addParticipant: (participant: ParticipantInVideoCall) => {
         set((state) => ({ participants: [...state.participants, participant] }));
     },
     setStreamForParticipant(stream: MediaStream, socketId: string, isShareScreen: boolean) {
@@ -62,13 +62,7 @@ export const useParticipantVideoCallStore = create<VideoCallState>()((set) => ({
         set((state) => ({ peerShareScreen: state.peerShareScreen.filter((p) => p.id != id) }));
     },
     clearPeerShareScreen: () => {
-        set((state) => {
-            state.peerShareScreen.forEach((p: any) => {
-                if(!p?.peer) return;
-                p?.peer?.destroy();
-            });
-            return { peerShareScreen: [] };
-        });
+        set((state) => ({ peerShareScreen: []}));
     },
     updatePeerParticipant: (peer: any, socketId: string) => {
         set((state) => ({
